+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod app;
 pub mod config;
 pub mod event;
+pub mod export;
 pub mod index;
 pub mod link;
 pub mod markdown;